Update patch set 2
Patch Set 2:

The approach looks ok (from what I know of package system) but it looks like a lot is being done in pkgman instead of package kit. I think eventually we would want the same functionality in SoftwareUpdater, so having the core of it in package kit would help for that. But I guess we can always refactor it later.

I agree "dist-upgrade" isn't really an appropriate name. Maybe something like local-update or offline-update? It doesn't even have to be from a specific media, you could have acquired the packages in any way you want and then use this to install, so I don't think "media" should be in the name of the command?

Patch-set: 2
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>