Add component maintainers info to ReadMe.md
This commit adds a maintainers table to ReadMe.md.

This is part of a marketing initiative to get more contributors, as well
as simplify workload and create a division of labour. Partly inspired by
the mailing list post from PulkoMandy here [1].

This marketing initiative consists of three stages:

1. Create a strong Haiku use case and compare with Linux and BSD. Use
Haiku forums and mailing list to gather ideas and turn them into a blog post.
2. Make a call for contributors, encourage people to become a Haiku developer.
Use Twitter, Reddit, HN etc. channels.
3. Purge the general opinion that Haiku is a hobbyist operating system, and
show people that we mean business, and we aim to create a viable Linux and
macOS competitior for Desktop systems. We should strive to increase general OS
usefulness, and create a stable system usable by everyone.

With this specific change we aim a few specific things:

1. Instead of telling newcomers just to go to Trac and select a ticket, present
an up-to-date table that they will feel comfortable to dive into, and encourage
them to keep maintaining that component. Links to the Trac component tickets
are provided for easy browsing.
2. Instead of rotating the already-minimal workforce, encourage people to work
on a specific component, and reduce possible burnout caused by the high number
of tickets and heavy workload.
3. Keep Trac contact and notification system up-to-date, clear obsolete
maintainer info. It doesn't make sense to keep inactive names, and would let
us know if a developer decides to stop contributing.  It will also help some
of the current developers working on areas currently not associated in Trac.
4. With newcomers, allow veteran developers to design a bigger vision for
Haiku, and help them spend more time on reviewing patches and plan future
development.

With the Gerrit post, a message to the haiku-development mailing list will be
sent. I believe we should try to fill the positions with the available
contributors (active/inactive), and encourage non-active developers to become
more active. When we think everyone available responds, we can proceed with
the second part of the marketing plan, aka, reach general public with our
cases, and ask for more help.

Constructive feedback welcome.

[1] https://www.freelists.org/post/haiku-development/Update-on-Haiku-R1-Release-Proposal,10

Change-Id: I9a90ed3b967162b217ecaf5fd5e5336cec333e70