Update patch set 1
Patch Set 1: Code-Review-2

Haiku does not have "component maintainers". We have people who are more or less knowledgeable about given components and do more or less work on them, but that is all; they are not "responsible" for the component in an official way, they do not have veto power over other non-"maintainer" changes, they do not control the official direction of the component. The default assignees in Trac are not "maintainers" and should not be considered such. When people have questions about a component, the default assignee in Trac may well reply to a ticket and answer them, but people should not email them directly (in fact when people contact me privately about XHCI, FreeBSD networking, or something else I have done a lot of work on, I almost always tell them to ask again publicly, and then I reply there: so that there is a record of what it was that was said, so that others can chime in if they have more relevant knowledge than I do, and so that the history is preserved for future reference.)

I have previously reiterated on all this a number of times with you on IRC, and I think previously on the mailing list and the like, about how we do not have "maintainers". So why have you gone ahead with this? At most, you should have started yet another mailing list discussion on the subject; Gerrit is not the place to have discussions about the organization of Haiku as a project.

Patch-set: 1
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-2