Update patch set 1
Patch Set 1:

> But ah, yes, status quo where no one touches the R1 tickets from 2 kilometres away is definitely preferable, yes indeed. Everyone is indeed rotating around the tickets, not within.

Huh? Where in the world are you drawing this from?

I spent a significant chunk of time whittling down the R1 milestone from >2000 tickets to well below 1500, and PulkoMandy has now gotten it all the way down to <700; not to mention all of the tickets we fixed from that milestone over the past few years instead of just shuffling things around. (For whatever it's worth, I was also the one to originally pull the trigger to change "Unscheduled" to be the default milestone instead of "R1", so that it quit filling up.)

> Responsibility in the sense of fixing bugs, yes. Thus, I believe it's an appropriate term in this context.

It is already generally acknowledged that people need to spend some time fixing bugs in a component instead of just adding new features all the time; in fact, part of the long dispute related to the Media Kit was about this. We hardly need to set "maintainers" to have this acknowledged and even "enforced".

> Nothing is stopping developers take on more than one component, and rotate between them. Or submit bug fixes for another component that already has a maintainer, or coordinate the work between maintainers.

Then I ask again: what exactly does "maintainer" imply, and why are you pushing for it so hard?

Patch-set: 1
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>