Update patch set 1
Patch Set 1:

> There are fixes that are happening, which I am grateful for all the developers that put their time into, but they seem to be mostly for the tickets from the "Unscheduled" milestone, which does not help the R1 milestone tickets, which are apparently required for the R1 release to happen. There are talks about the R1 release since pre-beta2, but no visible progress.

People fix the bugs they think need to be fixed, often because they are most complained about or they are affecting the person in question, and only rarely because they are merely part of some milestone. That is how Haiku has succeeded as a project: if we scratch our own itches first, we will have the patience to scratch everyone else's itches. If we stop doing that, we will probably all lose our motivation, as again, none of us are paid to do this.

> > It is already generally acknowledged that people need to spend some time fixing bugs in a component instead of just adding new features all the time ... we hardly need to set "maintainers" to have this acknowledged and even "enforced".
> 
> Irrelevant.

You cannot dismiss concerns as "irrelevant" without at least stating a reason why. This is a dialogue, you are not in a position of authority to simply end lines of debate with a single word.

> > Then I ask again: what exactly does "maintainer" imply, and why are you pushing for it so hard?
> 
> This is actually a very minor change that does not imply any secret agenda. It only tries to make it more newcomer friendly, and help them pick a component to start from. 

You keep asserting that this is somehow "more newcomer friendly". Do you have any evidence for this, e.g. people who did not start contributing because somehow they could not find who or where to ask questions, or other projects where people did get engaged because of such a list, or something like that?

> So then I ask now: Why do you oppose such a trivial change?

If it is such a trivial change, why do you advocate so strongly for it?

> As you told me, you are going to stop contributing to Haiku in the near future. Why shouldn't someone pick up where you're gonna leave off?

Yes, I may switch gears and start contributing to something else that I feel is presently more important, but as I have said before, if I had time for both, I would continue contributing to Haiku.

Of course someone should pick up where I left off, but what has this got to do with that? I already picked up from where a lot of other people left off (FreeBSD networking, XHCI, HaikuPorts, etc. etc. etc. that I am known for contributing a lot to were not my own creations!), and the same is true of PulkoMandy and most other contributors to the project. Somehow we have all picked up from where others left off.

Patch-set: 1
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>