Update patch set 1
Patch Set 1:

> > > Then I ask again: what exactly does "maintainer" imply, and why are you pushing for it so hard?
> > 
> > This is actually a very minor change that does not imply any secret agenda. It only tries to make it more newcomer friendly, and help them pick a component to start from. 
> 
> You keep asserting that this is somehow "more newcomer friendly". Do you have any evidence for this, e.g. people who did not start contributing because somehow they could not find who or where to ask questions, or other projects where people did get engaged because of such a list, or something like that?

This is just nitpicking, you know as well as I do that this is not something you can measure or put into statistics or present evidence. Nevertheless it is progress, in the incremental sense.

> > So then I ask now: Why do you oppose such a trivial change?
> 
> If it is such a trivial change, why do you advocate so strongly for it?

Because it is a progress.

> > As you told me, you are going to stop contributing to Haiku in the near future. Why shouldn't someone pick up where you're gonna leave off?
> 
> Yes, I may switch gears and start contributing to something else that I feel is presently more important, but as I have said before, if I had time for both, I would continue contributing to Haiku.
> 
> Of course someone should pick up where I left off, but what has this got to do with that? I already picked up from where a lot of other people left off (FreeBSD networking, XHCI, HaikuPorts, etc. etc. etc. that I am known for contributing a lot to were not my own creations!), and the same is true of PulkoMandy and most other contributors to the project. Somehow we have all picked up from where others left off.

Yes, someone will eventually pick up. The issue is, why not enable more means to make it sooner.

I am not going to go forward with this conversation further, because it is not helping with the points indicated in the commit message. From what I've understood here and your replies on IRC, you are just being emotional, and feeling disturbed from being presented ideas from a "non-technical person", as you've quoted (which is not correct BTW). I'd still think otherwise if you hadn't given your notice to the project (in a sense).

Patch-set: 1
Reviewer: Gerrit User 1000169 <1000169@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>