Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> This isn't a new or unique idea for the project.
> 
> Haiku historically (since the beginning) had folks "taking ownership" of sections of code. Early on this worked well.
> 
> However multiple times in the past this simply resulted in component maintainers shutting down from accepting constructive criticism of other developers.  We have had multiple instances of people feeling like they "own" parts of Haiku, resulting in a lot of conflict and strife in the community.
> 
> Thus, I'm a -2 on this one.   People are free to become experts in subject matter, but designating a hard "component maintainer" has historically resulted in bad results for us.

I think context really matters here. In my opinion, the concept of ownership has to do with the near-term goal of getting ready for R1. My proposal about ownership was a response against the pushback on the mailing list I got around using a time-based framework to get out the release. Instead component-owners would have the duty of reviewing the state of their component and be able to say go/no go for the readiness on their part of the system.

I agree that having powerful owners that leverage control over what goes in and does not go in, and about the future direction of their component is obviously not in the spirit of Haiku and it is not what we want to go for.

Obviously I do not fully know what Emir's vision for component owners is, but I just want to put out that there is another interpretation of ownership.

As for this patch, I appreciate the gesture, but I would be against codifying the ownership in the permanent history in Git, as I feel that these lists should be maintained in dynamic tools (like Trac).

Patch-set: 1
CC: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>