Update patch set 1
Patch Set 1: Code-Review-2

Hi,

I didn't want to be once again the first to reply and be under the impression that my opinion is biasing the debate. But I see there is some agreement from other devs, so I may share my view now.

So, here is how I feel lately. There has been much effort from Nielx and Emir to try to get an R1 release out. I appreciate it, but I think this, for me, has been counterproductive.

After a few evenings staring at the R1 bug list and streesing myself over 'I should be working on this' and not getting anything done, I have decided that this is not healthy and not motivating for me. So, I will not be part of it. Despite what you say in the commit message, for me, Haiku indeed is mainly a fun hobby project. Something I enjoy tinkering with when I'm back from work.

My situation at work currently has not been all that great. With the Covid-19 and other unrelated events, at the start of the year I was working on a nice project in a team of 40. That project has been stopped, I've been moved to a new customer on a short term work (about 2 months) and will be moved to yet another thing at the end of the month. Some of my colleagues are in temporary unemployment since march, waiting for a new project to start. It is already quite stressful for me, and clearly not the best time to be focusing on things I don't want to work on and be pressured in being more productive in Haiku as well.

So, I do not want to be a "maintainer" here for any component related to R1. I feel that I only get in the way of getting things done (it may or may not be the case, but that's how all the pressure makes me feel).

If you want to make Haiku more than a fun hobby for developers, create a company, hire some of them, and tell them what to do. Or use Haiku inc for that purpose if you manage to get Haiku inc to do something. Then you can maybe attempt to use more traditional management techniques to drive that workforce (and even then, I'm not totally sure it would work. Several of us developers are known to be rather the "free electron" kind of persons even in that situation, I think).

So, if you ever do this, please do not include me as a maintainer for anything.

And as nielx mentionned, it makes no sense to duplicate this in the README when the info is already in Trac.

Now, what are the productive things to do here?

- Improving our onboarding. Yes, telling people "oh, just pick a ticket in Trac and start hacking" maybe isn't great. It's hard to know if people actually leave because of this, if we never hear of them. We can think about having a per-component thing in Trac. It may be as simple as adding a "ticket status by component" to the milestone pages (as Trac own bugtracker does: https://trac.edgewall.org/milestone/1.5.2 ). This does help seeing where there is work to be done
- Giving more immediate feedback in Gerrit and optimizing our code review process: at the minimum, a build bot that is able to check changes submitted to Gerrit, making sure that they don't break the build, and maybe running haiku-format on them to detect the most obvious coding style problems. This provides some immediate feedback to patch submitters and I think it would help a lot. We can also try to make better use of assigning pending patches to developers (automatically or manually) so that we can get our list of pending patches down. It is quite frustrating to have patches lost in pages and pages of pending things and never reviewed and merged. And it can be confusing for someone who wants to continue working on something: which commit do they start from? What's the status of the work? How do they even know there is already work in progress?
- Get Haiku inc up and running and actively trying to hire someone for full-time work, if we are serious about getting things done.

There is no need to deeply change how the developers are organized. We work as a team and all take care of the whole OS. This is required, because our availability is variable (depending on other things we do, like, you know, having weekends) and so is our motivation to work on Haiku (vs. many other projects at least in my case). So, that's the thing that works best: depending on what's needed at a given time, someone will have the knowledge and motivation to dig into it. We can commit to do that once or maybe twice a year for releases in a more focused way, but other than that, we need some time to breathe (again, remember this is, for all of us, a fun hobby project we're not getting paid for).

Don't make this look like a job, because it's unpaid, and when you offer a job to people without paying them, they just leave. And rightfully so.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-2