Update patch set 1
Patch Set 1:

Hi Adrien,

Sorry to hear that things have not been that good with the COVID-situation. I hope you manage to keep your head up high and we will get through it!

I think we reached a stalemate here about our vision on how to create progress within the project and the community.

Your the standpoint boils down to that at this time we cannot release anything that can be marked as 'final'. The main argument behind it is that the software is not what we call R1 quality. The issue I keep running into with that argument is that (1) there is no reasonable clear standard of what R1 quality is and (2) whenever I try to come up with proposals to change this (like the time based release plan or using maintainers/owners as quality evaluators) you shoot it down.

I cannot conclude anything other than that your position is that the project/community should no longer aspire to releasing R1 and instead to embrace the hobby-developer nature of the project. And instead of setting Emir and me up as the ones trying to "change" the aspirations of the project, this looks to me like the real focus-shift.

In any case, I think arguing for pursuing a strategy for stable regular releases is moot at this point, because as you point out, there is not enough man power with the willingness to go for that goal to deliver.

N>

Patch-set: 1