Update patch set 1
Patch Set 1:

> Your the standpoint boils down to that at this time we cannot release anything that can be marked as 'final'. The main argument behind it is that the software is not what we call R1 quality. The issue I keep running into with that argument is that (1) there is no reasonable clear standard of what R1 quality is and (2) whenever I try to come up with proposals to change this (like the time based release plan or using maintainers/owners as quality evaluators) you shoot it down.

The R1 milestone in Trac currently reflects what I think R1 quality means. And you can't say I didn't make compromises here, after removing about a third of the tickets there.

If you want to ignore this work on defining the release that is the accumulated work of all contributors in managing the bugtracker over the years and trying to define what R1 is, yes, we won't find an agreement.

I don't think I have been shooting down things. I am challenging them and offering alternatives. I have said what I tink the plan should be already: manage to make yearly releases, or every 6 months or something, and work on this using agile methods: have a backlog, and at every release/sprint, decide what gets into the next thing we ship, and what doesn't. I have also said that I don't believe version numbers (calling the next release R1 or beta3 or whatever) won't help with our situation, and that R1 is not about quality and what tickets get in it, but rather in our ability to provide support for it as a stable release. If its lifetime is 6 months and after that we're back to telling people to use the nightlies, I won't consider it an R1, even if that's how its named.

> I cannot conclude anything other than that your position is that the project/community should no longer aspire to releasing R1 and instead to embrace the hobby-developer nature of the project. And instead of setting Emir and me up as the ones trying to "change" the aspirations of the project, this looks to me like the real focus-shift.

Which I why I said "if you do something like this, please do not register me as a maintainer for anything". That's all. You can ignore me, I am not in a specific position of power here, and in fact, anything that shows that I'm not in charge of Haiku is healthy for me. Please take it and bring it somewhere. Find new developers. I am clearly part of the "old grumpy guys" team now so I'm counter productive. I will work on other things where I can be of more use.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>