Update patch set 3
Patch Set 3:

By the way, if you use "Haiku" you need to take care of distro compatibility rules and have an alternative for de-branded build ("boot to system"? or something). Or at least use #error so people trying to build a de-trademarked version know they have to fix this and submit a patch.

Sorry for being annoying with non-essential details.

Patch-set: 3
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>