Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> The approach looks ok (from what I know of package system) but it looks like a lot is being done in pkgman instead of package kit. I think eventually we would want the same functionality in SoftwareUpdater, so having the core of it in package kit would help for that. But I guess we can always refactor it later.

I think that that could be done later. I think at this stage, these operations would be very specific to two use cases:
1. Users that have native systems that do not connect to the network.
2. Users of systems in locations that have poor broadband connectivity.

I think the first class of users will probably be on the side of 'enthusiasts' and will be open for some experimentation. The other class of users might benefit from a graphic SoftwareUpdater, however I think there are bigger fish to fry when it comes to getting in touch with those users.

> I agree "dist-upgrade" isn't really an appropriate name. Maybe something like local-update or offline-update? It doesn't even have to be from a specific media, you could have acquired the packages in any way you want and then use this to install, so I don't think "media" should be in the name of the command?

For me, the name should communicate that it is not just updating from any local directory or repository, but one that is specifically laid out like a Haiku Installer image, hence the requirement for having a `system/packages` and an optional `_packages_` directory, as well as repository configs for `Haiku` and `HaikuPorts`. 

The word medium does not cover all cases, but the word image is too vague and technically incorrect.

I am inclined to go for offline-upgrade (instead of your suggestion of offline-update).

Patch-set: 2
Reviewer: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>