Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > The approach looks ok (from what I know of package system) but it looks like a lot is being done in pkgman instead of package kit. I think eventually we would want the same functionality in SoftwareUpdater, so having the core of it in package kit would help for that. But I guess we can always refactor it later.
> 
> I think that that could be done later. I think at this stage, these operations would be very specific to two use cases:
> 1. Users that have native systems that do not connect to the network.
> 2. Users of systems in locations that have poor broadband connectivity.
> 
> I think the first class of users will probably be on the side of 'enthusiasts' and will be open for some experimentation. The other class of users might benefit from a graphic SoftwareUpdater, however I think there are bigger fish to fry when it comes to getting in touch with those users.

There can be many reasons to use offline upgrades. Limited internet access due to local wiring constraints or commercial practice of operators, filtered internet access (country/company/ISP/... limitation), lack of driver for the machine network devices, etc.

In general I think that Haiku is a "GUI-first" system and I'm uncomfortable making an exception for the package management. Why, with package management, suddenly it would be ok to have some things only available from the command line?

Sure, we can do it later, but since we have this in mind laready, why not already put the code in the right place in package kit to prepare for it? Instead of working on code we know will have to be heavily reworked when moving it there?

> For me, the name should communicate that it is not just updating from any local directory or repository, but one that is specifically laid out like a Haiku Installer image, hence the requirement for having a `system/packages` and an optional `_packages_` directory, as well as repository configs for `Haiku` and `HaikuPorts`. 
> 
> The word medium does not cover all cases, but the word image is too vague and technically incorrect.

Ideally, the code would handle both cases and detect what it is fed with. But that is something we can definitely do later :)

> 
> I am inclined to go for offline-upgrade (instead of your suggestion of offline-update).

Well, currently there is no notion of "upgrade" in pkgman. There is update (that will only install newer version of packages and their dependencies, and never remove anything or install lower versions) and there is full-sync (which is allowed to install lower versions and possibly remove a package to install another instead). Note that full-synccan be used to perform both an "upgrade" (to a newer release) or a "downgrade" (to an older release). It will just sync with whatever is in the repos.

We probably want something like full-sync here if this is to be used for updating from one release to another. Now I'm thinking that the easiest way to do this would be something like "pkgman full-sync /Haiku-DVD" (where /Haiku-DVD is the media to get the files from). Doing it this way we can also have "pkgman update /Haiku-DVD" if we want to (it can be useful if the user has already updated parts of the system/libs by other means and only wants the newer things from the DVD). And we don't need to add any more words.

Patch-set: 2