sd/mmc: read, naive method
First implementation of reading sectors from an SD card.
This is not the best performance for many reasons:
- No DMA
- Reads only one sector at a time
- Cannot read more than 512 bytes per syscall

Also there are major limitations:
- Cannot read less than 512 bytes. The hardware of course works in full
  sectors. The mmc_disk driver should go through the io scheduler to
  make sure requests have a reasonable size and offset, and nothing
  tries to read just a few bytes in the middle of a sector.
- SD cards only (no SDHC, no MMC)

Architecture problems:
- I think too much of the implementation is done in sdhci_pci and should
be moved to the upper layers. However it is difficult to say without
having implemented reads for multiple sectors (which is a higher level
thing, it needs a sequence of multiple commands) and DMA (which indeed
will be at the low level of the sdhci controller). It doesn't help that
the order of operations is a bit different depending on wether there is
DMA or not. In DMA mode you first prepare the buffer, then run the
command. In non-DMA mode you first send the command, then read the data
into the buffer. We need an API at the mmc_bus level that doesn't care
about that low-level detail. There are other things that the MMC bus
should be doing however, such as switching to different clock speeds
depending on which card is activated and how fast it can go.
- Not sure about the threading and locking in all this. I think at
either the mmc_bus or sdhci level, we should make sure that we don't
attempt to run two commands at the same time from different threads? Can
that happen? Or is that already handled somewhere in the device manager?

Change-Id: I688b6c694561074535c9c0c2545f06dc04b06e7d