Update patch set 7
Patch Set 7:

> By the way, we can use this opportunity to improve the body text as well. Current text does sound very weird to me, especially the part where it iterates selecting the keyboard layout again and again. Also text parts do not have clear transitions, it just cuts with brief statements. IMO, there is no need to ask a question as well.

I think the reason to mention that language and keymap can be changed later, too, is to not convey that you set those now and are stuck with it for this installation. I don't think asking what to do and providing the two buttons as answers is bad.
I do agree we could stream line the wording a bit. See Patch 8 for a suggestion.

Patch-set: 7
Reviewer: Gerrit User 1000004 <1000004@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>