Update patch set 8
Patch Set 8:

> Patch Set 7:
> I think the reason to mention that language and keymap can be changed later, too, is to not convey that you set those now and are stuck with it for this installation. I don't think asking what to do and providing the two buttons as answers is bad.
> I do agree we could stream line the wording a bit. See Patch 8 for a suggestion.

I thought it is a bit obvious that it can be changed afterwards, since it's a basic thing. And in theory they wouldn't be able to understand it anyway without selecting the correct language (if they are not bilingual). As for the layout, even if they do select the wrong layout, there's not much to type during installation.

And if I'm not mistaken when we exit Installer, it returns back to FirstBootPrompt, according to Panos' latest fixes, no? Thus, they can select another language and layout, not get stuck.

Patch-set: 8