Update patch set 8
Patch Set 8:

> And if I'm not mistaken when we exit Installer, it returns back to FirstBootPrompt, according to Panos' latest fixes, no? Thus, they can select another language and layout, not get stuck.

Yes, I think that makes the explanation redundant now. The remaining risk is getting to the desktop in a language you don't understand and having to find your way to preferences -> locale from there. Or you can always reboot and try again if you mis-click.

Now this text doesn't seem really useful since it's hard to not understand what's going on with the window (with all the icons, and after all there are only 2 buttons, a listview and a popup menu). It's there more to test if the selected language is indeed working, I guess :) (some languages have a message about the translation being a work in progress, for example)

Patch-set: 8