Update patch set 8
Patch Set 8:

> Patch Set 8:

> I thought it is a bit obvious that it can be changed afterwards, since it's a basic thing. And in theory they wouldn't be able to understand it anyway without selecting the correct language (if they are not bilingual). As for the layout, even if they do select the wrong layout, there's not much to type during installation.

I had a feeling that this will turn out to be another massive bikeshed...

It's true, we could do away with practically the complete text. After all, a list of languages and the icons should give a strong enough hint for anyone.
But a bit of a welcoming text, and assurance that you're NOT picking a single language/keymap combo now for the installation, is not a bad idea, IMO.

Patch-set: 8
Reviewer: Gerrit User 1000004 <1000004@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>