Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > And if I'm not mistaken when we exit Installer, it returns back to FirstBootPrompt, according to Panos' latest fixes, no? Thus, they can select another language and layout, not get stuck.
> 
> Yes, I think that makes the explanation redundant now. The remaining risk is getting to the desktop in a language you don't understand and having to find your way to preferences -> locale from there. Or you can always reboot and try again if you mis-click.

That risk is always gonna be, if the user so über-careless that they'd neglect selecting their language. In order to minimise the risk, Apple had a solution: https://make.wordpress.org/core/files/2014/05/osx-language.jpg

Maybe we could add the keyboard layout selection below this, and the next step would just be a dialog box, greeting the user and asking whether they want to try Haiku, or install it.

Patch-set: 8