Update patch set 8
Patch Set 8:

> That risk is always gonna be, if the user so über-careless that they'd neglect selecting their language. In order to minimise the risk, Apple had a solution: https://make.wordpress.org/core/files/2014/05/osx-language.jpg
> 
> Maybe we could add the keyboard layout selection below this, and the next step would just be a dialog box, greeting the user and asking whether they want to try Haiku, or install it.

That does not seem to be a solution to me. You can still accidentally click the button without selecting a language (or selecting the wrong one). And there is no text at all to confirm which language you selected.

What we have now is better. We can just remove the useless info from the text. And that's off-topic for this change anyway, so let's finish and merge this and see about it later :)

Patch-set: 8