Update patch set 8
Patch Set 8:

> That does not seem to be a solution to me. You can still accidentally click the button without selecting a language (or selecting the wrong one). And there is no text at all to confirm which language you selected.
> 
> What we have now is better. We can just remove the useless info from the text. And that's off-topic for this change anyway, so let's finish and merge this and see about it later :)

I agree, it's off topic. But still, the only real solution would be a neural link to the user's brain, but Apple's solution is more expressive, and uses more screen estate, thus getting more attention from the user. I bet it would trigger the need to scroll and find the appropriate language more than ours (not that ours is bad).

Patch-set: 8