Update patch set 8
Patch Set 8:

> That small one looks so cute!

Standalone, it does, but on a 1080p or more display, it makes a small window lost in the middle of the screen.

The current window is designed to just fit in 800x480, so it fits in small screens, but stays reeasonably large in high resolution ones. Difficult balance :)

Patch-set: 8