Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2:
> 
> This new patchset changes the translated strings, I don't know if this is desirable

I think it's better to have the whole help message as a single string to translate? It allows the translators to keep track of how these lines relate to each other (alignment, etc).

It's annoying that the message will have to be re-translated, but eventually it is for the better since in the long term it makes the translator's life easier.

Patch-set: 2